Scope chunked-corpus fixtures to tests/log/conftest.py
