+++ tests/conftest.py
@@ -5,14 +5,10 @@ import shutil
 import subprocess
 import sys
 import warnings
-from typing import TYPE_CHECKING
 
 import boto3
 import pytest
 from moto.server import ThreadedMotoServer
-
-if TYPE_CHECKING:
-    from test_helpers.chunked_corpus import ChunkedCorpus
 
 sys.path.append(os.path.join(os.path.dirname(__file__), "helpers"))
 
@@ -67,41 +63,6 @@ def pytest_addoption(parser):
 @pytest.fixture(scope="session")
 def local_inspect_tools(request):
     return request.config.getoption("--local-inspect-tools")
-
-
-# Chunked-format corpora (large-samples effort): converted once per
-# session; imports are function-local so conftest stays light for runs
-# that never request them.
-
-
-@pytest.fixture(scope="session")
-def chunked_corpus(tmp_path_factory: pytest.TempPathFactory) -> "ChunkedCorpus":
-    """Chunked conversions of every test `.eval` log (default chunk size).
-
-    Realistic writer-policy corpus: most samples fit a single chunk.
-    """
-    from test_helpers.chunked_corpus import build_chunked_corpus
-
-    from inspect_ai.log._recorders.chunked.format import DEFAULT_CHUNK_SIZE
-
-    return build_chunked_corpus(
-        tmp_path_factory.mktemp("chunked_corpus"), DEFAULT_CHUNK_SIZE
-    )
-
-
-@pytest.fixture(scope="session")
-def chunked_corpus_small_chunks(
-    tmp_path_factory: pytest.TempPathFactory,
-) -> "ChunkedCorpus":
-    """Chunked conversions with a tiny chunk size (multi-chunk samples)."""
-    from test_helpers.chunked_corpus import (
-        CORPUS_SMALL_CHUNK_SIZE,
-        build_chunked_corpus,
-    )
-
-    return build_chunked_corpus(
-        tmp_path_factory.mktemp("chunked_corpus_small"), CORPUS_SMALL_CHUNK_SIZE
-    )
 
 
 @pytest.fixture(autouse=True)
